Remove some redundant code and fix a missing import
